+++ Samples~/SquareGridMesh/MeshGeneration/Generators/SquareGridGenerator.cs
@@ -73,36 +73,35 @@ namespace MeshGeneration.Generators
             }
         }
     }
+    
+    [Serializable]
+    public struct GridResolution
+    {
+        [Min(1)] public int x;
+        [Min(1)] public int y;
+    }
 
     [CreateAssetMenu(menuName = "ScriptableObjects/MeshGeneration/Generators/SharedSquareGridGenerator"
                    , fileName = "SharedSquareGridGenerator")]
-    public class SquareGridGenerator : MeshGenerator
+    public class SquareGridGenerator : MeshGenerator<GridResolution>
     {
-        [Serializable]
-        public struct GridResolution
-        {
-            [Min(1)] public int x;
-            [Min(1)] public int y;
-        }
-
-        [field: SerializeField] public GridResolution Resolution    { get; set; }
-        int                                           NumCells      => Resolution.x       * Resolution.y;
-        private            int                        TriangleCount => NumCells           * 2;
-        protected override int                        VertexCount   => (Resolution.x + 1) * (Resolution.y + 1);
-        protected override int                        IndexCount    => 6                  * NumCells;
+        private            int NumCells      => Data.x       * Data.y;
+        private            int TriangleCount => NumCells       * 2;
+        protected override int VertexCount   => (Data.x + 1) * (Data.y + 1);
+        protected override int IndexCount    => 6              * NumCells;
 
         public override void CalculateVertices()
         {
             Vertices          = new NativeArray<Vertex>(VertexCount, Allocator.TempJob);
-            VerticesJob       = new CalculateVerticesJob {vertices = Vertices, size = 1, resolution = int2(Resolution.x, Resolution.y)};
-            VerticesJobHandle = ((CalculateVerticesJob) VerticesJob).ScheduleParallel(Resolution.y + 1, 1, default);
+            VerticesJob       = new CalculateVerticesJob {vertices = Vertices, size = 1, resolution = int2(Data.x, Data.y)};
+            VerticesJobHandle = ((CalculateVerticesJob) VerticesJob).ScheduleParallel(Data.y + 1, 1, default);
         }
 
         public override void CalculateTriangles()
         {
             Triangles          = new NativeArray<int3>(TriangleCount, Allocator.TempJob);
-            TrianglesJob       = new CalculateTrianglesJob {triangles = Triangles, resolution = int2(Resolution.x, Resolution.y)};
-            TrianglesJobHandle = ((CalculateTrianglesJob) TrianglesJob).ScheduleParallel(Resolution.y, 1, default);
+            TrianglesJob       = new CalculateTrianglesJob {triangles = Triangles, resolution = int2(Data.x, Data.y)};
+            TrianglesJobHandle = ((CalculateTrianglesJob) TrianglesJob).ScheduleParallel(Data.y, 1, default);
         }
         
         public override void FinishMeshCalculations()